Face Recognition Python Project
My project in face recognition using Python utilizes the power of computer vision to detect and recognize faces in images and video streams. To accomplish this, I am utilizing the OpenCV library, which provides pre-trained models for face detection as well as the capability to train custom models using a dataset of labeled images. The project involves capturing video from a webcam or other device, processing the video to detect faces, and then comparing the detected faces to a database of known individuals to determine the identity of the person in the video.

Furthermore, I have also implemented additional features such as facial landmark detection and facial expression recognition to enhance the overall performance of the model. To increase the accuracy and robustness of the model, I have also trained the model with a custom dataset.

Overall, the goal of this project is to provide real-time face recognition capabilities for various applications such as security systems, human-computer interaction, and many more. With the implementation of advanced techniques and custom dataset, this project aims to achieve high accuracy and robustness in recognizing faces in real-world scenarios.